refactor: implement createShareBlob function for image and icon composition in ShareButton
